add talent stats as separate component
